+++ src/mappings/control/events/memberRemoved.ts
@@ -1,4 +1,4 @@
-import { Context, EventItem, Block } from '../../../processor'
+import { Context, Block } from '../../../processor'
 import { Event } from '../../../types/generated/support'
 
 import { getMemberRemovedData } from './getters'
@@ -8,14 +8,14 @@ import { arrayToHexString, addressCodec } from '../../util/helpers'
 import { ObjectNotExistsWarn } from '../../../common/errors'
 
 
-async function handleMemberRemovedEvent(ctx: Context, block: Block, item: EventItem) {
-	const eventData = getMemberRemovedData(ctx, item.event)
+async function handleMemberRemovedEvent(ctx: Context, block: Block, event: Event, name: string) {
+	const eventData = getMemberRemovedData(ctx, event)
 	let address = addressCodec.encode(eventData.who)
 	let orgId = arrayToHexString(eventData.orgId)
 
 	let member = await getOrgMember(ctx.store, orgId, address)
 	if (!member) {
-		ctx.log.warn(ObjectNotExistsWarn(item.name, 'Member', `${orgId}-${address}`.toLowerCase()))
+		ctx.log.warn(ObjectNotExistsWarn(name, 'Member', `${orgId}-${address}`.toLowerCase()))
 		return
 	}
