+++ src/mappings/signal/index.ts
@@ -1,0 +1,20 @@
+import { Pallet } from '../types/pallet'
+import { handleProposalCreatedEvent } from './events/proposalCreated'
+import { handleProposalVotedEvent } from './events/proposalVoted'
+import { handleProposalEvent } from './events/proposal'
+
+
+export default {
+	name: 'Signal',
+	callHandlers: {},
+	eventHandlers: {
+		'Created': handleProposalCreatedEvent,
+		'Voted': handleProposalVotedEvent,
+		'Activated': handleProposalEvent,
+		'Accepted': handleProposalEvent,
+		'Rejected': handleProposalEvent,
+		'Expired': handleProposalEvent,
+		'Aborted': handleProposalEvent,
+		'Finalized': handleProposalEvent,
+	}
+} as Pallet;
