--- temp/flow/events/campaignFailed.ts
@@ -1,31 +1,0 @@
-export { Event, Block } from '../../../types/generated/support'
-import { Context, EventItem } from '../../../processor'
-import { getCampaignFailedData } from './getters'
-
-import { getCampaign } from '../../util/db/getters'
-import { storage } from '../../../storage'
-
-import { arrayToHexString } from '../../util/helpers'
-import { ObjectNotExistsWarn, StorageNotExistsWarn } from '../../../common/errors'
-
-
-async function handleCampaignFailedEvent(ctx: EventHandlerContext, block: SubstrateBlock, event: Event) {
-	const campaignIdArray = getCampaignFailedData(ctx, event)
-	let campaignId = arrayToHexString(campaignIdArray)
-
-	let campaign = await getCampaign(ctx.store, campaignId)
-	if (!campaign) {
-		ctx.log.warn(ObjectNotExistsWarn('Campaign', campaignId))
-		return
-	}
-	const stateStorageData = await storage.control.getOrgStateStorageData(ctx, block, campaignIdArray)
-    if (!stateStorageData) {
-		ctx.log.warn(StorageNotExistsWarn(event.name, campaignId))
-		return
-    }
-
-	campaign.state = stateStorageData.__kind
-	await ctx.store.save(campaign)
-}
-
-export { handleCampaignFailedEvent }
