--- temp/sense/index.ts
@@ -1,12 +1,0 @@
-import { Pallet } from '../types/pallet';
-import { handleEntityCreatedEvent } from './events/entityCreated';
-import { handlePropertyUpdatedEvent } from './events/propertyUpdated';
-
-export default {
-	name: 'Sense',
-	extrinsicHandlers: {},
-	eventHandlers: {
-		'EntityCreated': handleEntityCreatedEvent,
-		'PropertyUpdated': handlePropertyUpdatedEvent,
-	}
-} as Pallet;
